10-05-2020
se agrega la parte de edificio, se modifica el codigo para que el server de la db ingrese la fecha y hora y no por medio de codigo, aun falta validar los cambios

+++ estadoHistorico.php
@@ -39,13 +39,9 @@
     {
         $data = $mensaje->obtenerJSON();
 
-        if(isset($data['estado']))
+        if(isset($data['estado']) && is_string($data['estado']))
         {
-            $item = array(
-            'estado' =>$data['estado']
-        );
-
-            $api->add($item);
+            $api->add($data);
         }
         else
         {   
@@ -60,12 +56,7 @@
 
         if(isset($data['id']) && isset($data['estado']))
         {
-            $item = array
-            (
-                'id'     =>$data['id'],
-                'estado' =>$data['estado']
-            );
-            $api->update($item);
+            $api->update($data);
         }
         else
         {
@@ -82,7 +73,15 @@
         if(isset($data['id']))
         {
             $id = $data['id'];
-            $api->delete($id);
+
+            if(is_numeric($id))
+            {
+                $api->delete($id);
+            }
+            else
+            {
+                $mensaje->error('Los parametros son incorrectos'); 
+            }
         }
         else
         {
